Include current `$post` with `wp_enqueue_media()`
If editing a post (post type), this allows additional settings to
be enqueued for media handling. Specifically, this allows the auto
embed of content from a URL such as YouTube.

If editing a user, term, etc... `$post` will be `null`, which is
the default argument in `wp_enqueue_media()`, resulting in no change.