Book title doesn't update properly (fix #542)
The custom metadata plugin calls the  function: update_metadata() in the WP Core file: web/wp/wp-includes/meta.php.

Because the title is empty, i.e. not in the postmeta table, we end up in  `return add_metadata()`. before reaching the end of the update, I.e.

```
 $meta_ids = $wpdb->get_col( $wpdb->prepare( "SELECT $id_column FROM $table WHERE meta_key = %s AND $column = %d", $meta_key, $object_id ) );
 if ( empty( $meta_ids ) ) {
  return add_metadata( $meta_type, $object_id, $raw_meta_key, $passed_value );
 }
```

The fix is to use the `added_post_meta` hook in conjunction with the `updated_post_meta` hook.