Remove old `latex_path` option handling (#923)
This appears to be a leftover from WP LaTeX. In that plugin, the
`latex_path` option and the `AUTOMATTIC_LATEX_LATEX_PATH` constant
are used only if the `Automattic_Latex_DVIPNG` class is controlling
the LaTeX output. These are not used in the classed used to generate
LaTex via WPCOM, which Pressbook includes.

It also appears that in the fork of that plugin, the admin form
inputs that could have controlled the `latex_path` option value
were removed. This results in the value always being `false`, which
in turn results in `which latex` being run on the server every time
the `PBLatexAdmin` class is started.