allow other latex renderers to be used for pdf exports (#928)
When attempting a PDF export, if the selected latex renderer was
being loaded from outside of the pressbooks codebase, the export
would fail with an error saying that no such latex render file
exists. This solution uses the `pb_require_latex` filter hook to
reference/load the appropriate latex renderer if it lives outside
of the pressbooks codebase (e.g., another plugin).