Allow users to choose upload field for custom import choices
The current logic forces the user to user an input filed that
wants an http link to the file. But there are some situations
where users need to have the flexibility to choose what input
type they use for their custom import type.  This allows users
to use the `pb_import_table_cell` hook to render the input field
of choice. This also introduces an element id `pb-other` that
shows/hides the field depending on what import type is selected.