Add title attribute to section divs (fix PB-14549) (#1441)
This PR adds a `title` attribute to front matter, chapter and back matter `<div>` elements in XHTML exports. The `title` attribute is set to the front matter, chapter, or back matter's short title (if present) or the title, if the short title is not set. This allows the [CSS generated content](https://www.w3.org/TR/css-gcpm-3/#named-strings) `section-title` property to be consistently set from this attribute instead of trying to set it from one of two sources (the title element or a hidden short title element). When Buckram >= 1.2.0 (see corresponding PR), the hidden short title element will not be included as it will no longer be needed.