Only fetch revision IDs for status and visibility metabox (fix #1458) (#1465)
In the status and visibility metabox, we call `wp_get_post_revisions()` to:

1. Get a revision count for the current post.
2. Get the ID of the latest revision.

This call did not supply any arguments, so full post objects for each revision would be retrieved. Under normal circumstances this was not causing performance problems, but a client recently reported issues loading a 43,000 word post with ~160 revisions. This PR changes the `wp_get_post_revisions()` call to retrieve revision IDs only, which should resolve these issues.