Don't import deleted posts (WXR import) (#1471)
Currently, a Pressbooks or WordPress WXR file with deleted posts in it will list them alongside published posts on the import selection screen. This can lead to the importing of unexpected content (e.g. an older version of a post which has been trashed), or, at best, the inability to distinguish between a deleted post and a published post with the same name. This PR ignores posts with status = `trash` when listing posts for import. An alternative approach could be to label these as `Post Name (Deleted)` but that seems like it could be a source of confusion.