Use png to jpeg as failsafe routine (#1474)
This PR does two things. It allows the Exception to be an exception under all exceptional circumstances...rather, it removes the potential for a `PHP Fatal` if `$pid` is not able to return error codes (if the file is empty) 

![image](https://user-images.githubusercontent.com/2048170/48921711-802f9400-ee56-11e8-8d86-fb5a3666e344.png)

The certain circumstances alluded to above in which the file will return empty is captured here: https://stackoverflow.com/questions/18719256/why-does-pdftoppm-poppler-utils-have-no-jpeg-option 

Under the `pdftoppm` has no `-jpeg` argument, the error code from the first command will be something other than `0` and no file will be created. 

![screen shot 2018-11-21 at 4 39 45 pm](https://user-images.githubusercontent.com/2048170/48921850-72c6d980-ee57-11e8-8baa-f4d6dd83b166.png)

Under those two conditions the proposed PR will convert to png then use `imagemagick` to convert png to jpg.