Fancy quoted URL (Fix #1493) (#1498)
The media shortcode in the test document is:

`[media src=”https://www.youtube.com/watch?v=Lqqsp8soXTo” /]`

The first thing that catches my eye is that the quote being used `”` (aka fancy quote) is not the same as `"` 

Same thing with `’` vs `'`

When editing this test file in LibreOffice, fancy quotes seem to be on lock down. It is difficult for me to get the correct quotes in there. I can sympathize with this pain point.

## Workaround:

However, If I remove the quotes it works:

`[media src=https://www.youtube.com/watch?v=Lqqsp8soXTo /]`

## Fix:

In PHP The URL is comming in as: `”https://www.youtube.com/watch?v=Lqqsp8soXTo”`

If I trim fancy quotes before validating the URL, the shortcode works.