Chapter license tweaks (fix #1508) (#1563)
The previous PR missed a case where if a chapter had an empty (unset) license, it would still show the chapter's title and link to the chapter in the license statement. As per #1508, the expected behaviour here would be to show the book's title and link to the book in the license statement.