SSE (Event Stream) Forms (fix #1282) (#1569)
* Playing with generators and event streams

* More info durring progress

* The fun never stops!

* Working code!

* Better user feedback

Main change: $_SESSIONS are read only when in AJAX so we needed another way to store pb-notices.

Other things too.

* Better counters

* Fix parts counter

* Add eventsource.polyfill

Because it doesn't work with Microsoft Edge

https://wpdev.uservoice.com/forums/257854-microsoft-edge-developer/suggestions/6263825-server-sent-events-eventsource

https://developer.mozilla.org/en-US/docs/Web/API/EventSource#Browser_compatibility

* Make the JS that catches form more generic

* Tests for EventStreams class

* Add Utility\PercentageYield class

* Example XHTML Export refactoring

* Proof of concept.

Buffer inside a buffer not working... :(

* Fix for nested buffers

Buffers are nested. While one buffer is active, flushes from child buffers are not really sent to the output, but rather to the parent buffer. Only when there is no parent buffer are contents sent to the browser.

* Example using Epub exports

* Typo

* Refactor if/else order

* Simplify

Placeholder for PHP 7.4 typed properties.
(`public array $msgStack;`)

Adds ability to add messages.

Etc.

* Show export errors, show reload link on eventstream fails

* JS Lint

* Emit error if no export format was selected

* count up timer

* Stop clock on error

* Clock clock

* Turn off heartbeat where things take a long time

Heartbeat script will sometimes return an error (connection timeout, or similar) when PHP is busy ignoring it doing something else (exporting, importing, cloning)

If the browser gets this kind og error, the EventEmitter dies.

* Warn user about leaving page

* Code comments for Ned

* Import JS

* Status bar on import page

* Import examples (Epub & API)

* Cover generator, sessions

* Fix cache (too fast?)

* Fix linting errors