Fix PHP Notices when cloning H5P
Cloning triggers 'the_content' but we don't use it. This caused PHP Notices with H5P because the shortcodes aren't fixed until the end of the process.

WordPress API let's us specify which fields we care about in `WP_REST_Controller::get_fields_for_response` when we do an inrternal `rest_do_request`

I've set it to `id` where possible.