Fix PHP Notices (E_NOTICE) when cloning H5P (#1612)
Cloning triggers 'the_content' but we don't use it. This caused PHP Notices with H5P because the shortcodes aren't fixed until the end of the process.

WordPress API let's us specify which fields we care about in `WP_REST_Controller::get_fields_for_response` when we do an internal `rest_do_request`

I've set it to `id` where possible.