Fix clone & import timeouts (fix #1673) (#1700)
## cURL error 28: Operation timed out

This means the server that the book is hosted on is very slow, or doing some sort of bandwidth throttling. I  increased timeout, switched from  http version 1.0 to 1.1, and set the compress flag to true.

## EventStream Connection Error

Because there's a maximum $_GET length, and our form often exceeds it, we can't send ?url=parameters directly to EventSource. The workaround is to submit using jQuery Form Plugin, set a transient, callback EventSource on done, pick up where we left off.