https://github.com/pressbooks/pressbooks-book/issues/588 Chapter title does not always override book title in chapter-level license statement when an chapter-level author has been added #588
I checked the code, the original code requires that if the book-level license is the same as the chapter-level license, then use the book-level license.
Which is in conflict with the issue 588. Since the issue 588 looks more logic, I change the code to reflect the change. The maintainor needs to verify the
product definition if the issue 588 reportor did not do so.