Remove /wp-json/pressbooks/v2/books/search
This is not good enough. We need to rethink how to do this. Possible hint: WP_Site_Query, WP_Network_Query, "Now after the filter `networks_pre_query` and `sites_pre_query run`, the code will exit straight away. This allows developers to completely hot-wire the network and site query, to load from another source, such as a different cache or Elastic search."

Source: https://make.wordpress.org/core/2019/10/15/enhancements-to-the-network-sites-screen-in-wordpress-5-3/